requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
seaborn
fpdf2
openpyxl
python-dateutil